+++ Project/Src/Utility/Utility.cpp
@@ -294,14 +294,14 @@ VECTOR Utility::Normalize(const VECTOR& v)
 
 Vector2 Utility::Normalize(const Vector2& v)
 {
-    return v / Magnitude(v);
+    return v / (float)Magnitude(v);
 }
 
 Vector2 Utility::Normalize(const Vector2I& v)
 {
      if (v.x == 0 && v.y == 0) { return Vector2(0, 0); }
 
-    return Vector2(v / sqrtf(v.x * v.x + v.y * v.y));
+     return Vector2(v / sqrtf((float)(v.x * v.x + v.y * v.y)));
 }
 
 VECTOR Utility::Clamp(const VECTOR& v, const VECTOR& minV, const VECTOR& maxV)
